+++ source/isaaclab/isaaclab/terrains/terrain_generator_cfg.py
@@ -140,10 +140,10 @@ class TerrainGeneratorCfg:
 
     border_below_ground: bool = True
     """Whether to place the border below the ground. Defaults to True.
-    
-    If True, the heighest point of the border is at the height 0.0. This extends the terrain and is typically used in 
+
+    If True, the highest point of the border is at the height 0.0. This extends the terrain and is typically used in
     locomotion tasks.
-    If False, the lowest point of the border is at the height 0.0. This is typically used in navigation tasks and acts 
+    If False, the lowest point of the border is at the height 0.0. This is typically used in navigation tasks and acts
     as a wall to prevent the robot from falling off the terrain.
     """